+++ machines/tom/services/ollama/default.nix
@@ -3,6 +3,7 @@
 {
   services.ollama = {
     enable = true;
+    loadModels = [ "gemma4:26b" ];
     package = pkgs.ollama-cuda;
     models = "/etc/ollama/models";
     user = "ollama";